+++ src/main/java/com/github/onsdigital/dpjavadurationparse/ParseDuration.java
@@ -1,4 +1,4 @@
-package parse;
+package com.github.onsdigital.dpjavadurationparse;
 
 import java.time.Duration;
 import java.time.format.DateTimeParseException;
@@ -8,17 +8,22 @@ import java.util.Map;
 public class ParseDuration {
 
     /**
-     * Checks whether duration starts with string looks like an ISO8601 formatted string, and if not it adds PT.
+     * Parse's Duration in either ISO8601 or format's similar to: 3h56s345ms.
+     *
+     * The "PT" prefix on each of the ISO8601's durations example below, where P is
+     * a "period" duration designation ("P") and a "time" designation ("T") per the ISO-8601Checks whether duration
+     * starts with string looks like an ISO8601 formatted string, and if not it adds PT.
+     *
      * @param stringDuration Checking whether {@link String} starts with the appropriate ISO8601 characters.
      * @return A {@link Duration}
      * Compatible with one or a combination of the following units: h, m, s
      *
      * Example Valid Formats
-     *          PT3h45m56s345ms34us
-     *          -PT3h45m56s345ms34µs
-     *          PT3h56s345ms34µs
-     *          -PT3h56s34µs
-     *          PT3h34µs
+     *          PT3h45m56s345ms
+     *          -PT3h45m56s345ms
+     *          PT3h56s345ms
+     *          -PT3h56s
+     *          PT3h
      *          -PT56s
      *
      *          3h45m56s345ms34us
@@ -34,7 +39,7 @@ public class ParseDuration {
      *          3h56s345ms34µs*
      */
     /*
-        The "PT" prefix on each of the Duration's string representations shown above indicates that the representation is a "period" duration designation ("P") and a "time" designation ("T") per the ISO-8601
+
      */
     public static Duration parseThisDuration(String stringDuration) {
         // Flag to identify durations in the past
@@ -66,43 +71,5 @@ public class ParseDuration {
             return null;
         }
         return parsedDuration;
-    }
-
-
-    public static void main(String[] args) {
-        Map<String,Duration> validDurations = new HashMap<>();
-
-//        System.out.println("parseDuration(\"PT3h45m56s345ms34us\"); = " + parseDuration("PT3h45m56s345ms34us"));
-//        System.out.println("parseDuration(\"PT3h45m56s345ms34us\"); = " + parseDuration("PT3h45m56s345ms"));
-//        System.out.println("parseDuration(\"PT3h45m56s345ms34us\"); = " + parseDuration("PT3h45m50s"));
-//        System.out.println("parseDuration(\"PT3h45m56s345ms34us\"); = " + parseDuration("3h45m50s"));
-//        System.out.println("parseDuration(\"PT3h45m56s345ms34us\"); = " + parseDuration("3h45m50s"));
-//        System.out.println("parseDuration(\"PT3h45m56s345ms34us\"); = " + parseDuration("-3h45m50s"));
-        System.out.println("parseDuration(\"PT3h45m56s345ms34us\"); = " + parseThisDuration("-3h45m50s"));
-
-//        "PT3h45m56s345ms34us"
-//        "-PT3h45m56s345ms34µs"
-//        "PT3h56s345ms34µs"
-//        "-PT3h56s34µs"
-//        "PT3h34µs"
-//        "-PT56s"
-//
-//    if (parseDuration())
-//
-//
-//  "3h45m56s345ms34us"
-//             "-3h45m56s345ms34µs"
-//             "3h56s345ms34µs"
-//             "-3h56s34µs"
-//             "3h34µs"
-//             "-56s"
-//
-//
-//        this.turn("P1DT1H10M10.5S");
-//        Duration.in(DAYS, HOURS, MINUTES).between(start, end); // equivalent alternative for Java 8
-
-//        Duration.parsePeriod("-P18DT2H"); // -18 days and -2 hours (negative)
-//        Duration<CalendarUnit> duration2 = Duration.parseCalendarPeriod("P0000-11-04"); // 11 months and 4 days
-
     }
 }
